Final Files
This is a python GUI based game which allow multiple players to play
this game by changing their turn.
Whenever a box is formed score of that player increased by one.
Algorithm: Everytime whenever a line formed whether a box is formed or
not. To check this we used a Tkinter function to see into a double size
square which have 2 corners in both side of line with equal distance and
two vertex along the line.
Also whenever box formed we check if all box are formed then declare
winner and game is over then.